Add files via upload
This script is a process monitor that logs information about new processes created on a Windows system. It uses the WMI interface to watch for new process creation events and logs information such as the time of creation, user, executable path, command line, process ID, parent process ID, and privileges enabled. The script also defines functions to obtain the list of privileges enabled for a process and log messages to a file. The script runs in an infinite loop and logs information about new processes as they are created.